refactor(infra): created function to generate the rebalancing config reliably
